refactor(cell): move Magnetism from source_estate to source_cell
The Magnetism class only depends on source_base and is held by value as a
member of UnitCell (source_cell/unitcell.h). It is a fundamental property of
the unit cell, so it belongs in source_cell rather than source_estate.

This removes the unitcell.h -> source_estate header dependency (a core data
structure no longer reaches up into the electronic-state module) and breaks
one direction of the source_cell <-> source_estate cycle.

- git mv source_estate/magnetism.{h,cpp} -> source_cell/
- update all includers to source_cell/magnetism.h
- move the source file entry between CMakeLists and fix the unit-test path
