+++ source/source_cell/magnetism.cpp
