+++ source/source_cell/magnetism.h
