+++ source/source_estate/module_charge/charge.cpp
@@ -27,7 +27,7 @@
 #include "source_base/timer.h"
 #include "source_base/tool_threading.h"
 #include "source_cell/unitcell.h"
-#include "source_estate/magnetism.h"
+#include "source_cell/magnetism.h"
 #include "source_hamilt/module_xc/xc_functional.h"
 #include "source_io/module_parameter/parameter.h"
