+++ source/source_lcao/module_dftu/dftu.cpp
@@ -6,7 +6,7 @@
 #include "source_base/inverse_matrix.h"
 #include "source_base/memory_recorder.h"
 #include "source_base/timer.h"
-#include "source_estate/magnetism.h"
+#include "source_cell/magnetism.h"
 #include "source_estate/module_charge/charge.h"
 
 #include <cmath>
